Factored The Tile class into a better OO form
Before the Tile class would store the suit as a variable.

The issue with this is that, we now have to ask a lot of if's

If we instead inherit for the suit, then it's a bit better

If there were more suits I'd probably mix them in?

This is an open question for myself