Started Shanten calculation
Started Shanten calculations.

This has caused a fair bit of changes and bugs found.

Namely `valueSort` had a bug where if you gave it 345 and 66 the values would align, this is fixed by bit shifting the answer

Further the code has some views to make viewing things better

A future commit will remove `acceptComplete` as the shanten accept is superior