live up to the real spirit of examples
Some of my old code would give back a graph which is less useful to compose off of.

Now that we have more advanced views, let us use the tiles directly and link to the given view on the data for showing. No need to do it before hand like we were doing.