Create a git dependnecy graph to track parent and child relations
The old method was OK (after fixing the bug in the last commit it went from 10 seconds to 1.9!), however it is slow and lacks the ability to create more queries.

Using an explicit graph we can do more relations, and I've tracked both children and parents so we can do more refined queries in the futre.

Overall the design is higher quality, although requiring more code.


There is one bug however, the `learnFinalizedNode:` improperly overally constrains what nodes can have relations. and is thus commented out in the `computeRelationsWith:` method.


If this is ever fixed we should enable it again.