Add a basic github query series of classes
Before this was contained within the projectboard itself. This is not ideal, as the querying code is now tied to the projectboard class which is harder to instantitate than we could other classes, making followup queries harder.

This does not currently replace any of that, but it gets us most of the way there, we even recreate the "What's cooking view".

In the future the State should be computed from some kind of enumerable class