Incorporated the query context into the projectboard
This is a sensible change, as instead of holding three slots to be able to query informaiton, we made a dedicated class for this.

What this means is that the API is simpler and more generic as this same information can be derived for other kinds of queries that are more specific.