cache DepGraph computation and load it lazily
Before the depGraph would try to compute upon the inspection, causing a lag on inspecting GitProjectBoards, now it does so a bit more lazily.

Further we also cache the loading of the depGraph for easier inspection use, and easier updating.

We also use the standard items for it so we aren't having to pass around a closure.

Sadly this code doesn't fix the issue where without clicking on any views, the forwarding view would automatically run an update on the nodes, causing polling to happen in the background, even when unwanted.