Create the initial work for having Issues work
The code used to assume everything was a PR. Now we've added rudimentary support for Issues.

Issues work on the whatscooking view now, and are properly filtered from the graph view.

Sadly we have to check fields to determine the class, as __typename is currently bugged in GT.

In the future I'll need to extend this to WIP issues.

Before then I need to invert who has the labels field, as WIP Labels lack tags.