Overhauled Projct Structure
I tried to keep this commit somewhat minimal, but all the changes I
did broke the code in some way, so they are all squashed together
here.

We tried to remove redundnacy in items, by having a clear separations
in the query and the data.

Before the data was spread over with redundant information, now it's
much better.

A query class has the class it fills up inside, as it queries into
that for the most part.

We did a bunch of changes:

1. We made queriers for repos, users, and tooling boards much more
useful
  - Before the queriers for project boards would be tied to
  organizations only, however it's now liberated to be both users and
  organizations
2. Project Boards have expanded, working on issues
3. ProjectBoardItems and it's subclasses massively changed
   - They now contain query objects, no longer items

There are other changes we should do in the future as well.
1. Being able to convert more classes to their query classes
2. new examples for issues