Make the git tooling work with unknown topics
We extend the tool not to error when topics are unknown.

Instead they will show a different color depending on the context.

This is greatly needed as nothing is worse than finding out you can't use the tool as you forgot to add an extra remote from a new contributor